# frontend/light-task-frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kantano",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/assistant/conversations": {
      "get": {
        "tags": [
          "Assistant"
        ],
        "summary": "List Conversations",
        "operationId": "list_conversations_api_projects__project_id__assistant_conversations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationRead"
                  },
                  "title": "Response List Conversations Api Projects  Project Id  Assistant Conversations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Assistant"
        ],
        "summary": "Create Conversation",
        "operationId": "create_conversation_api_projects__project_id__assistant_conversations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/assistant/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "Assistant"
        ],
        "summary": "Get Conversation",
        "operationId": "get_conversation_api_projects__project_id__assistant_conversations__conversation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Assistant"
        ],
        "summary": "Delete Conversation",
        "operationId": "delete_conversation_api_projects__project_id__assistant_conversations__conversation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/assistant/conversations/{conversation_id}/runs": {
      "post": {
        "tags": [
          "Assistant"
        ],
        "summary": "Start Run",
        "operationId": "start_run_api_projects__project_id__assistant_conversations__conversation_id__runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/assistant/conversations/{conversation_id}/runs/{run_id}/decision": {
      "post": {
        "tags": [
          "Assistant"
        ],
        "summary": "Decide Action",
        "operationId": "decide_action_api_projects__project_id__assistant_conversations__conversation_id__runs__run_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionCall": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "start": {
            "type": "integer",
            "title": "Start"
          },
          "end": {
            "type": "integer",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "id",
          "start",
          "end"
        ],
        "title": "ActionCall"
      },
      "ActionDecision": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          },
          "actionId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Actionid"
          }
        },
        "type": "object",
        "required": [
          "approve",
          "actionId"
        ],
        "title": "ActionDecision"
      },
      "ActionDisplay": {
        "properties": {
          "steps": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ActionStep"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "column_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column Id"
          },
          "new_column_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Column Id"
          },
          "before_column_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Column Id"
          },
          "tag_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Id"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "tag_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Ids"
          }
        },
        "type": "object",
        "title": "ActionDisplay"
      },
      "ActionOutcome": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "rejected",
              "failed",
              "skipped"
            ],
            "title": "Status",
            "default": "completed"
          },
          "tool": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "CreateTask",
                  "UpdateTask",
                  "MoveTask",
                  "CreateColumn",
                  "RenameColumn",
                  "MoveColumn",
                  "CreateTag",
                  "UpdateTag",
                  "AddTagToTask",
                  "RemoveTagFromTask"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          },
          "action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Id"
          },
          "step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Id"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "created",
                  "updated",
                  "moved",
                  "column_created",
                  "column_renamed",
                  "column_moved",
                  "tag_created",
                  "tag_updated",
                  "tag_added_to_task",
                  "tag_removed_from_task"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "column_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column Id"
          },
          "tag_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "type": "object",
        "title": "ActionOutcome"
      },
      "ActionStep": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tool": {
            "type": "string",
            "enum": [
              "CreateTask",
              "UpdateTask",
              "MoveTask",
              "CreateColumn",
              "RenameColumn",
              "MoveColumn",
              "CreateTag",
              "UpdateTag",
              "AddTagToTask",
              "RemoveTagFromTask"
            ],
            "title": "Tool"
          },
          "args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Args"
          },
          "display": {
            "additionalProperties": true,
            "type": "object",
            "title": "Display"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tool",
          "args"
        ],
        "title": "ActionStep"
      },
...
      "ChatDetail": {
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationRead"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageRead"
            },
            "type": "array",
            "title": "Messages"
          },
          "latestRun": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "conversation",
          "messages",
          "latestRun"
        ],
        "title": "ChatDetail"
      },
...
      "ConversationCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "default": "Новый чат"
          }
        },
        "type": "object",
        "title": "ConversationCreate"
      },
      "ConversationRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "projectId": {
            "type": "integer",
            "title": "Projectid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "enum": [
              "local",
              "cloud"
            ],
            "title": "Mode"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "projectId",
          "title",
          "mode",
          "createdAt",
          "updatedAt"
        ],
        "title": "ConversationRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageCreate": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MessageCreate"
      },
      "MessageRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/TaskReference"
            },
            "type": "array",
            "title": "References"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "references",
          "createdAt"
        ],
        "title": "MessageRead"
      },
...
      "ProposedAction": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "CreateTask",
              "UpdateTask",
              "MoveTask",
              "CreateColumn",
              "RenameColumn",
              "MoveColumn",
              "CreateTag",
              "UpdateTag",
              "AddTagToTask",
              "RemoveTagFromTask",
              "ExecutePlan"
            ],
            "title": "Name"
          },
          "args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Args"
          },
          "display": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionDisplay"
              },
              {
                "type": "null"
              }
            ]
          },
          "action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Id"
          },
          "tool_call_id": {
            "type": "string",
            "title": "Tool Call Id"
          },
          "calls": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ActionCall"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calls"
          }
        },
        "type": "object",
        "required": [
          "name",
          "args",
          "tool_call_id"
        ],
        "title": "ProposedAction"
      },
...
      "RunRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "pending",
              "executing",
              "completed",
              "rejected",
              "failed",
              "interrupted",
              "unknown"
            ],
            "title": "Status"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "stepCount": {
            "type": "integer",
            "title": "Stepcount"
          },
          "fallbackUsed": {
            "type": "boolean",
            "title": "Fallbackused"
          },
          "proposedAction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposedAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "provider",
          "model",
          "stepCount",
          "fallbackUsed",
          "proposedAction",
          "result",
          "createdAt",
          "updatedAt"
        ],
        "title": "RunRead"
      },
      "RunResult": {
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionOutcome"
            },
            "type": "array",
            "title": "Actions"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "RunResult"
      },
...
      "TaskReference": {
        "properties": {
          "type": {
            "type": "string",
            "const": "task",
            "title": "Type",
            "default": "task"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "TaskReference"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}